requirements.txt
torch
torchvision
streamlit
scikit-learn
scikit-image
numpy
pillow